+++ tests/index.test.ts
@@ -1,4 +1,5 @@
 import './access.test'
+import './matchSource.test'
 import './noteTemplate.test'
 import './openAlexMapper.test'
 import './zoteroResponse.test'